backend/pyproject.toml
[project]
name = "o2c-email-agent"
version = "0.1.0"
description = "FastAPI backend for O2C AR email processing."
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.116.0",
  "langchain-anthropic>=0.3.0",
  "langchain-core>=0.3.0",
  "langchain-google-genai>=2.1.0",
  "langchain-groq>=0.3.0",
  "langchain-ollama>=0.2.0",
  "langchain-openai>=0.3.0",
  "langgraph>=0.3.0",
  "opentelemetry-exporter-otlp-proto-grpc>=1.32.0",
  "opentelemetry-instrumentation-fastapi>=0.53b0",
  "opentelemetry-sdk>=1.32.0",
  "psycopg[binary]>=3.2.0",
  "prometheus-client>=0.22.0",
  "pydantic-settings>=2.10.0",
  "sqlalchemy>=2.0.39",
  "uvicorn>=0.35.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.5",
  "ruff>=0.11.0",
]

[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[tool.uv]
package = true

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
